fixed type of pop

+++ src/main/java/report/friction/models/weather/DailyWeather.java
@@ -42,6 +42,6 @@ public class DailyWeather extends Weather {
     @JsonProperty("temp")
     private Map<String, Double> temperature;
     @JsonProperty("pop")
-    private Integer probabilityOfPrecipitation;
+    private Double probabilityOfPrecipitation;
 
 }
